Move token config into its own file
