+++ js/main.mjs
@@ -8,6 +8,7 @@ import * as audio from "./audio.mjs";
 import * as controls from "./controls.mjs";
 import * as dialog from "./dialog.mjs";
 import * as placeables from "./placeables/index.mjs";
+import * as tokenConfig from "./token-config.mjs";
 import * as spritesheets from "./spritesheets.mjs";
 import * as scripts from "./scripts.mjs";
 import * as pixelate from "./pixelate.mjs";
@@ -29,6 +30,7 @@ Hooks.on("init", ()=>{
     controls,
     dialog,
     placeables,
+    tokenConfig,
     spritesheets,
     scripts,
     pixelate,
